changed the typedef for the two 32-bit int types from long to int.  long is 8 bytes on Mac OS X 10.8, but 4 bytes on Windows i686 and x86_64 (!!)

+++ lib/sdif.h
@@ -91,8 +91,8 @@ Music and Audio Technologies, University of California, Berkeley.
 
     typedef unsigned short	sdif_unicode;
     typedef short		sdif_int16;
-    typedef long		sdif_int32;
-    typedef unsigned long	sdif_uint32;
+    typedef int		sdif_int32;
+    typedef unsigned int	sdif_uint32;
     typedef float		sdif_float32;
     typedef double		sdif_float64;
